requirements.txt
alembic==1.18.4
bcrypt==5.0.0
blinker==1.9.0
click==8.3.3
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.5.0
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.12
MarkupSafe==3.0.3
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
